Improve project structure and Split App into Components
