--- src/App.module.css
@@ -1,35 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.red {
-  color: red;
-}
-
-.bold {
-  font-weight: bold;
-}
-
-.App button {
-  background-color: green;
-  color: white;
-  font: inherit;
-  border: 1px solid pink;
-  padding: 8px;
-  cursor: pointer;
-}
-
-.App button:hover {
-  background-color: lightgreen;
-  color: black;
-}
-
-.App button.Red {
-  background-color: red;
-
-}
-
-.App button.Red:hover {
-  background-color: salmon;
-  
-}
