+++ src/components/Cockpit/Cockpit.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import classes from './Cockpit.module.css'
+
+const cockpit = (props) => {
+    const assignedClasses = [];
+    let btnClass = '';
+    if (props.showPersons) {
+        btnClass = classes.Red;
+    }
+
+    if (props.persons.length <= 2) {
+        assignedClasses.push(classes.red) // classes = ['red']
+    }
+    if (props.persons.length <= 1) {
+        assignedClasses.push(classes.bold) // classes =['red', 'bold']
+    }
+
+    return (
+        <div className={classes.Cockpit}>
+            <h1> Hello world </h1>
+            <p className={assignedClasses.join(' ')}>This is working</p>
+            <button 
+                className={btnClass}
+                onClick = { props.clicked } > Toggled Persons 
+            </button>  
+        </div>
+
+    );
+};
+
+export default cockpit;
