+++ src/components/Cockpit/Cockpit.module.css
@@ -1,0 +1,30 @@
+.red {
+    color: red;
+}
+  
+.bold {
+    font-weight: bold;
+}
+  
+.Cockpit button {
+    background-color: green;
+    color: white;
+    font: inherit;
+    border: 1px solid pink;
+    padding: 8px;
+    cursor: pointer;
+}
+
+.Cockpit button:hover {
+    background-color: lightgreen;
+    color: black;
+}
+
+.Cockpit button.Red {
+    background-color: red;
+}
+
+.Cockpit button.Red:hover {
+    background-color: salmon;
+}
+
