+++ src/components/Persons/Person/Person.jsx
@@ -2,11 +2,6 @@ import React from 'react';
 import classes from './Person.module.css';
 
 const person = (props) => {
-    const rnd = Math.random();
-    if (rnd > 0.7) {
-        throw new Error ('Something went wrong');
-    }
-    
     return (
         <div className={classes.Person}>
             <p>I'm {props.name} Bear.</p>
