+++ src/containers/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import classes from './App.module.css';
-import Person from './Person/Person.jsx';
+import Persons from '../components/Persons/Persons';
+import Cockpit from '../components/Cockpit/Cockpit';
 
 class App extends Component {
     state = {
@@ -47,49 +48,22 @@ class App extends Component {
 
     render() {
         let persons = null;
-        let btnClass = '';
 
         if (this.state.showPersons) {
-            persons = ( 
-                <div> {
-                    this.state.persons.map((person, index) => {
-                        return <Person
-                            click = {
-                                () => this.deletePersonHandler(index)
-                            }
-                            name = { person.name }
-                            age = { person.age }
-                            key = { person.id }
-                            changed = {
-                                (event) => this.nameChangedHandler(event, person.id)
-                            }
-                        />
-                       
-                    })
-                } 
-                </div> 
-            );
-        }
-
-        btnClass = classes.Red;
-        const assignedClasses = [];
-        if (this.state.persons.length <= 2) {
-            assignedClasses.push(classes.red) // classes = ['red']
-        }
-        if (this.state.persons.length <=1) {
-            assignedClasses.push(classes.bold) // classes =['red', 'bold']
+            persons =  <Persons 
+                persons={this.state.persons}
+                clicked={this.deletePersonHandler}
+                changed={this.nameChangedHandler}/>;
         }
 
         return ( 
-            
             <div className = {classes.App}>
-            <h1> Hello world </h1>
-            <p className={assignedClasses.join(' ')}>This is working</p>
-            <button className={btnClass}
-            // onClick={this.switchNameHandler.bind(this, "Max")}>Switch name</button>
-            onClick = { this.togglePersonsHandler } > Toggled name 
-            </button> 
-            { persons }   
+                <Cockpit 
+                    showPersons={this.state.showPersons}
+                    persons={this.state.persons}
+                    clicked={this.togglePersonsHandler}
+                />
+                { persons }   
             </div>
         );
     }
